customer-service/requirements.txt
# customer-service — Django REST Framework + MySQL
Django==4.2.11
djangorestframework==3.15.1
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1